+++ controller/__init__.py
@@ -1,0 +1,8 @@
+from flask import Flask
+from dotenv import load_dotenv
+
+load_dotenv()
+
+app = Flask(__name__)
+
+import controller.controller
